June 1996 version
This is identified as MFE1 0.1 in the code comments and dates from
June 1996. I believe this is "F" code -- the short-lived modern
subset of Fortran 90 for which there was a compiler based on tooling
from the NAG compiler. The current source is for the Navier-Stokes
equation and the single mfein input file for the Sod shock problem.