June 1996 update
This reorganizes the files and adds problems for several different
PDE systems. Also lots of massaging of the source code of an
apparently cosmetic nature. This code is from a directory originally
named MFE1-0.1

The code was modified to use the nonstandard MCLOCK function
provided by the XLF compiler and won't compile today. To be fixed
in the next commit.