New character variable input procedure
This adds a specific procedure for the generic `read_tagged_data`
that reads a character valued variable. This code is originally
from the `1d/new` directory and dates from Feb 2000. Also added
to that directory was a new 4-stream boron diffusion problem
associated with my work at Motorola on stress-mediated dopant
diffusion. The setup of that problem was not completed however,
so I've opted not to include it here. The original sources for
that problem come from a Motorola backup CD in the directory
`Codes/SMD/msd`